requirements.txt
# ============================================================================
# REQUIREMENTS.TXT - AIPHA v0.0.3 & CGAlpha v0.0.1
# Generado: 1 de Febrero de 2026
# Solucionado: Problema P0 Crítica #1
# ============================================================================

# Core CLI & UI
click>=8.0.0
rich>=10.0.0

# Data Processing & Analysis
pandas>=1.3.0
numpy>=1.20.0
duckdb>=0.5.0

# Machine Learning & Models
scikit-learn>=0.24.0
joblib>=1.0.0

# Configuration & Validation
pydantic>=2.0.0
python-dotenv>=0.21.0

# System & Utilities
psutil>=5.8.0

# Infrastructure (Redis cache & queues)
redis>=5.0.0

# LLM Integration (CRITICAL - Solución P0 #2)
openai>=1.0.0
requests>=2.28.0

# Testing (Development)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code Craft Sage - Phase 3: Test Generator
jinja2>=3.0.0

# Code Craft Sage - Phase 4: Git Automator
gitpython>=3.1.0

# Optional: Performance & Debugging
python-dateutil>=2.8.0
